[auth] moved DataInitializer to dbCore
