+++ src/main/java/ru/otus/vsh/knb/dbCore/DataInitializer.java
@@ -1,4 +1,4 @@
-package ru.otus.vsh.knb.webCore.services;
+package ru.otus.vsh.knb.dbCore;
 
 public interface DataInitializer {
     void createInitialData();
